[PIDM-1763] change amount type from Integer to Long (pagopa-ecommerce-commons) (#259)
* feat(amount): update amount to long type from int

* feat(amount): update amount to long type from int

* feat(amount): update amount to long type from int

* feat(amount): update amount to long type from int

* feat(amount): update amount to long type from int

* fix: change 12000 to 12000L

* chore: set <version>3.3.1-rc</version>

* chore: testa

* [PIDM-1959] deploy snapshot-commons (#263)

* chore: test snap permission

* chore: add snapshot publish workflow on PR

* chore: add commit id suffix to snapshot version name

* chore: fix Check failure

* chore: fix Check failure permission

* chore: fix Check failure permission

* chore: update actions/checkout from v1 to v4.2.2

* chore: back v1

* chore: update actions/checkout version 6.0.2

* chore: resolve comment PR

* chore: add ACT guards for local testing with act

* chore: add snapshot guard to prevent accidental release deletion

* fix: improve error handling in cleanup snapshot action

* fix: set package_name it.pagopa.pagopa-ecommerce-commons

* fix: set package_name it.pagopa.pagopa-ecommerce-commons

* chore: remove redundant SNAPSHOT guard in cleanup-snapshot action

---------


* Bump version [skip ci]

* chore: trigger snapshot build

* chore: Remove Integer overload from euroCentsToEuro utility method

* fix: Remove duplicated test in EuroUtilsTest

---------

+++ src/main/java/it/pagopa/ecommerce/commons/documents/PaymentTransferInformation.java
@@ -15,7 +15,7 @@ import org.springframework.data.mongodb.core.mapping.Document;
 public class PaymentTransferInformation {
     private String paFiscalCode;
     private Boolean digitalStamp;
-    private Integer transferAmount;
+    private Long transferAmount;
     private String transferCategory;
 
     /**
@@ -29,7 +29,7 @@ public class PaymentTransferInformation {
     public PaymentTransferInformation(
             String paFiscalCode,
             Boolean digitalStamp,
-            Integer transferAmount,
+            Long transferAmount,
             String transferCategory
     ) {
         this.paFiscalCode = paFiscalCode;
